get controller configs working real-time end to end

+++ app/src/main/java/com/jeremyhahn/cropdroid/model/ChannelConfig.kt
@@ -1,5 +1,5 @@
 package com.jeremyhahn.cropdroid.model
 
-data class ChannelConfig(val id: Int, val name: String, val enable: String, val notify: String,
+data class ChannelConfig(val id: Long, val name: String, val enable: String, val notify: String,
                    val condition: String, val schedule: String, val duration: String,
                    val debounce: String, val backoff: String, val state: Int)
