+++ app/src/main/java/com/jeremyhahn/cropdroid/model/Schedule.kt
@@ -2,7 +2,7 @@ package com.jeremyhahn.cropdroid.model
 
 import java.util.*
 
-data class Schedule(val id: Int, var channelId: Int, var startDate: Calendar, var endDate: Calendar?, var frequency: Int,
+data class Schedule(val id: Long, var channelId: Int, var startDate: Calendar, var endDate: Calendar?, var frequency: Int,
         var interval: Int, var count: Int, var days: ArrayList<String>) {
 
         constructor() : this(0, 0, Calendar.getInstance(), null, 0, 0, 0, ArrayList())
