+++ app/src/main/java/com/jeremyhahn/cropdroid/ui/microcontroller/menu/ChannelTimerMenuItem.kt
@@ -19,7 +19,7 @@ import java.io.IOException
 class ChannelTimerMenuItem(context: Context, menu: ContextMenu, channel: Channel, cropDroidAPI: CropDroidAPI) {
 
     init {
-        menu!!.add(0, channel.id, 0, "Timer")
+        menu!!.add(0, channel.id.toInt(), 0, "Timer")
             .setOnMenuItemClickListener(MenuItem.OnMenuItemClickListener() {
                 val inflater: LayoutInflater = LayoutInflater.from(context)
