+++ app/src/main/java/com/jeremyhahn/cropdroid/ui/schedule/ScheduleListRecyclerAdapter.kt
@@ -134,7 +134,7 @@ class ScheduleListRecyclerAdapter(val activity: ScheduleListActivity, val cropDr
 
             Log.d("onCreateContextMenu", "schedule: " + schedule)
 
-            menu!!.add(0, schedule.id, 0, "Edit")
+            menu!!.add(0, schedule.id.toInt(), 0, "Edit")
                 .setOnMenuItemClickListener(MenuItem.OnMenuItemClickListener() {
 
                     val _adapter = this.adapter
@@ -184,7 +184,7 @@ class ScheduleListRecyclerAdapter(val activity: ScheduleListActivity, val cropDr
                     true
                 })
 
-            menu!!.add(0, schedule.id, 0, "Delete")
+            menu!!.add(0, schedule.id.toInt(), 0, "Delete")
                 .setOnMenuItemClickListener(MenuItem.OnMenuItemClickListener() {
                     activity.deleteSchedule(schedule)
                     true
